Refactored test_softdelete.
